+++ R/globals.R
@@ -1,4 +1,5 @@
 #'
 utils::globalVariables(c("name", "iso_a2", "iso_a3", "geometry",
                          "MapFiller", "X", "Y",
-                         "iso_a2_eh", "iso_a3_eh"))
+                         "iso_a2_eh", "iso_a3_eh",
+                         "map_df0"))