+++ tests/testthat/test2.R
@@ -1,7 +1,0 @@
-worldplot(data = testdata1b,
-          div = 1,
-          ColName = "VNum",
-          CountryName = "Cshort",
-          CountryNameType = "isoa2",
-          rangeVal = c(0,50),
-          annote = FALSE)