+++ tests/testthat/test5.R
@@ -1,7 +1,0 @@
-worldplot(data = testdata1b,
-          div = 1,
-          ColName = "VNum",
-          CountryName = "Cnames",
-          CountryNameType = "name",
-          rangeVal = c(0,50),
-          annote = FALSE)