+++ NEWS.md
@@ -1,3 +1,13 @@
+# WorldMap 1.2.0
+
+* map_df0 is now saved as one of the dataset, meaning that it is not built from scratch every time
+  a function is called. Because of this, rnaturalearth and rnaturalearthdata packages are not required to be imported anymore
+* Adjusted some label positions
+* Added the shadows option, to choose whether to print a shadow around country
+  labels or not
+* Improved the UK_as_GB option in countrycoord_data
+
+
 # WorldMapR 1.1.0
 
 * Added the internal function countrycoord_data, which substitutes geometries_data and incorporates